feat(rit): add logger and remove 3rd party dependencies + refactor
